+++ CHANGELOG.md
@@ -9,7 +9,15 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
-- `enum DataSourceKey`, which is used in fetching a data source from the server.
+- The server side should call `server::handle` to handle incoming requests from
+  a client. This function takes a handler, which should implement the `Handler`
+  trait. The handler currently handles the following request:
+  - `GetDataSource`
+- Types required in the requests handled by `Handler`:
+  - `DataSourceKey`, `DataType`, `DataSource`
+- The `client` module provides the following functions to send requests to the
+  server:
+  - `Connection::get_data_source`
 
 ## [0.8.0] - 2024-10-11